refactor(mcv): edit servises
issue #36
